1.0.1 Hotfix
- Fixed bug where the Node Content would still be the Clipboard on the GUI mode.
- Code now checks when a character is changed on the input fields instead of when you click on another field.
- CLI mode now warns you when you will overwrite a file.
- Fixed a small bug related to file naming on the CLI version
- Fixed File Signature check running where there is no Signature Parameter

+++ defs/new_node_file.py
@@ -8,14 +8,17 @@ import os
 current_format_version = chr(0x01)
 
 def new_file(node_name, node_key, node_address, node_content = pyperclip.paste()):
-    if not os.path.isdir("nodes"):
-        os.mkdir("nodes")
-
     # Hash is in cSHAKE256
     node_key_bytes = process_input(node_key, node_address)
 
+    # If Nodes folder does not exist, create one
+    if not os.path.isdir("nodes"):
+        os.mkdir("nodes")
+
+    # Generates a random cSHAKE256 hash if no node_name parameter was received
     if node_name == None:
-        node_name = process_input(f"{get_random_bytes(16)}", f"{get_random_bytes(16)}")
+        random_hash = process_input(f"{get_random_bytes(16)}", f"{get_random_bytes(16)}")
+        node_name = random_hash.hex()
 
     if node_content == None:
         print("Your Node Content/Clipboard is empty!")
@@ -26,6 +29,8 @@ def new_file(node_name, node_key, node_address, node_content = pyperclip.paste()
         print("Node Name: ", node_name)
         print(f"Node Content: {node_content}\n")
 
+    print("Generating Cipher... ", end="")
+
     cipher = AES.new(node_key_bytes, AES.MODE_GCM)
     encrypted_content, node_signature = cipher.encrypt_and_digest(bytes(node_content, "utf-8"))
 
@@ -35,14 +40,27 @@ def new_file(node_name, node_key, node_address, node_content = pyperclip.paste()
 
     print(f"AES Cipher Generated! Unique Signature: {node_signature.hex()}")
 
-    return node_name.hex(), cipher_nonce, cipher_text
+    return node_name, cipher_nonce, cipher_text
 
 
 
-def write_file(node_file_path, cipher_nonce, cipher_text):
+def write_file(node_file_path, cipher_nonce, cipher_text, cli_overwrite_warning):
     if(core.debug):
         print(f"\nFile Path Passed to write_file: {node_file_path}")
+
+    if cli_overwrite_warning:
+        if os.path.exists(node_file_path):
+            print("\nA Node with that name already exists! Are you sure you want to overwrite it? Type YES to confirm.")
+            
+            if input("Confirmation: ") != "YES":
+                print("\nNode Creation cancelled!")
+                return
+
+            # Small workaround ^^'
+            print("")
 
     file = open(node_file_path, "w")
     file.write(f"CRYPTODOT{current_format_version}\n{cipher_nonce}\n{cipher_text}")
     file.close()
+
+    print("Node Created!")
